Completed register and auth with scatum

+++ app/Http/Controllers/V1/ImageManipulationController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\UpdateImageManipulationRequest;
 use App\Http\Resources\V1\ImageManipulationResource;
 use App\Models\Album;
 use App\Models\ImageManipulation;
+use Illuminate\Http\Request;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Str;
@@ -21,13 +22,17 @@ class ImageManipulationController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        return ImageManipulationResource::Collection(ImageManipulation::paginate());
+        return ImageManipulationResource::Collection(ImageManipulation::where('user_id', $request->user()->id)->paginate());
     }
 
-    public function byAlbum(Album $album)
+    public function byAlbum(Request $request, Album $album)
     {
+        if($request->user()->id != $album->user_id)
+        {
+            return abort(403, 'Unauthorized');
+        }
         $where = [
             'album_id' => $album->id,
         ];
@@ -52,12 +57,16 @@ class ImageManipulationController extends Controller
         $data = [
             'type' => ImageManipulation::TYPE_RESIZE,
             'data' => json_encode($all),
-            'user_id' => null,
+            'user_id' => $request->user()->id,
         ];
 
         if(isset($all['album_id']))
         {
-            //TODO
+            $album = Album::find($all['album_id']);
+            if($request->user()->id != $album->user_id)
+            {
+                return abort(403, 'Unauthorized');
+            }
 
             $data['album_id'] = $all['album_id'];
         }
@@ -106,8 +115,12 @@ class ImageManipulationController extends Controller
      * @param  \App\Models\ImageManipulation  $image
      * @return \Illuminate\Http\Response
      */
-    public function show(ImageManipulation $image)
+    public function show(Request $request, ImageManipulation $image)
     {
+        if($request->user()->id != $album->user_id)
+        {
+            return abort(403, 'Unauthorized');
+        }
         return new ImageManipulationResource($image);
     }
 
@@ -117,8 +130,12 @@ class ImageManipulationController extends Controller
      * @param  \App\Models\ImageManipulation  $image
      * @return \Illuminate\Http\Response
      */
-    public function destroy(ImageManipulation $image)
+    public function destroy(Request $request, ImageManipulation $image)
     {
+        if($request->user()->id != $album->user_id)
+        {
+            return abort(403, 'Unauthorized');
+        }
         $image->delete();
         return response('', 204);
     }
